ml/requirements.txt
transformers
torch
datasets
boto3
python-dotenv
Pillow